tsql-core/src/test/java/com/xy/xsql/tsql/converter/functions/conversion/ParseConverterTest.java
package com.xy.xsql.tsql.converter.functions.conversion;

import com.xy.xsql.block.core.printer.MetaBlockPrinter;
import com.xy.xsql.block.model.BlockMeta;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2018/5/31.
 */
public class ParseConverterTest {

    @Test
    public void testMetaPrint() throws Exception {
        BlockMeta b = ParseConverter.meta;

        StringWriter writer = MetaBlockPrinter.print(b);

        System.out.println(writer);
        Assert.assertEquals(writer.toString(),
                "<PARSE> ::=\n" +
                        "PARSE ( string_value AS data_type [ USING culture ] )");
    }
}